move BoxCoder to mlperf helpers
